fix: host urls mismatch
