+++ apps/erp/package.json
@@ -129,6 +129,7 @@
   },
   "devDependencies": {
     "@carbon/config": "workspace:*",
+    "@carbon/dev": "workspace:*",
     "@lingui/macro": "catalog:",
     "@lingui/vite-plugin": "catalog:",
     "@react-router/dev": "catalog:",
