+++ apps/mes/app/routes/[.well-known]+/$.tsx
@@ -1,0 +1,5 @@
+import { data } from "react-router";
+
+export function loader() {
+  return data(null, { status: 404 });
+}
