+++ apps/mes/vite.config.ts
@@ -1,3 +1,4 @@
+import { applyDotenvToProcessEnv } from "@carbon/dev/vite";
 import { reactRouter } from "@react-router/dev/vite";
 import { lingui } from "@lingui/vite-plugin";
 import path from "node:path";
@@ -5,51 +6,55 @@ import { defineConfig, PluginOption } from "vite";
 import babelMacros from "vite-plugin-babel-macros";
 import tsconfigPaths from "vite-tsconfig-paths";
 
-export default defineConfig(({ mode, isSsrBuild }) => ({
-  build: {
-    minify: true,
-    rollupOptions: {
-      onwarn(warning, defaultHandler) {
-        if (warning.code === "SOURCEMAP_ERROR") {
-          return;
-        }
+export default defineConfig(({ mode, isSsrBuild }) => {
+  applyDotenvToProcessEnv(mode, __dirname);
 
-        defaultHandler(warning);
+  return {
+    build: {
+      minify: true,
+      rollupOptions: {
+        onwarn(warning, defaultHandler) {
+          if (warning.code === "SOURCEMAP_ERROR") {
+            return;
+          }
+
+          defaultHandler(warning);
+        },
+        ...(isSsrBuild && { input: "./server/app.ts" }),
       },
-      ...(isSsrBuild && { input: "./server/app.ts" }),
     },
-  },
-  define: {
-    global: "globalThis",
-  },
-  ssr: {
-    noExternal: [
-      "react-dropzone",
-      "react-icons",
-      "react-phone-number-input",
-      "tailwind-merge",
-    ],
-  },
-  server: {
-    port: 3001,
-    allowedHosts: [".ngrok-free.app", ".w.modal.host", ".w.modal.dev", ".dev", ".localhost"],
-  },
-  plugins: [
-    babelMacros(),
-    lingui(),
-    reactRouter(),
-    tsconfigPaths(),
-  ] as PluginOption[],
-  resolve: {
-    alias: {
-      "@carbon/utils": path.resolve(
-        __dirname,
-        "../../packages/utils/src/index.ts"
-      ),
-      "@carbon/form": path.resolve(
-        __dirname,
-        "../../packages/form/src/index.tsx"
-      ),
+    define: {
+      global: "globalThis",
     },
-  },
-}));
+    ssr: {
+      noExternal: [
+        "react-dropzone",
+        "react-icons",
+        "react-phone-number-input",
+        "tailwind-merge",
+      ],
+    },
+    server: {
+      port: 3001,
+      allowedHosts: [".ngrok-free.app", ".w.modal.host", ".w.modal.dev", ".dev", ".localhost"],
+    },
+    plugins: [
+      babelMacros(),
+      lingui(),
+      reactRouter(),
+      tsconfigPaths(),
+    ] as PluginOption[],
+    resolve: {
+      alias: {
+        "@carbon/utils": path.resolve(
+          __dirname,
+          "../../packages/utils/src/index.ts"
+        ),
+        "@carbon/form": path.resolve(
+          __dirname,
+          "../../packages/form/src/index.tsx"
+        ),
+      },
+    },
+  };
+});
