+++ packages/dev/vite.ts
@@ -1,0 +1,30 @@
+import path from "node:path";
+import { loadEnv } from "vite";
+
+/**
+ * Merge `.env*` files into `process.env` so SSR code that reads `process.env`
+ * (e.g. `@carbon/auth`, `@carbon/env`) sees the same values as Vite's
+ * `import.meta.env`.
+ *
+ * App-local files are loaded first, then repo-root files (last wins) so
+ * `crbn up`–written root `.env.local` overrides stale app-level copies.
+ *
+ * In non-production modes, file values **overwrite** existing `process.env`
+ * keys — `react-router dev` can invoke the vite config with modes other than
+ * `"development"` during startup, which previously left stale shell values
+ * (e.g. `SUPABASE_URL=127.0.0.1:54321`) in place.
+ */
+export function applyDotenvToProcessEnv(mode: string, appDir: string) {
+  const repoRoot = path.resolve(appDir, "../..");
+  const fromFiles = {
+    ...loadEnv(mode, appDir, ""),
+    ...loadEnv(mode, repoRoot, ""),
+  };
+  const devOverwrite = mode !== "production";
+  for (const [key, value] of Object.entries(fromFiles)) {
+    if (value === undefined || value === "") continue;
+    if (devOverwrite || process.env[key] === undefined) {
+      process.env[key] = value;
+    }
+  }
+}
